fix GetSporkValue return

+++ src/spork.cpp
@@ -101,7 +101,7 @@ bool IsSporkActive(int nSporkID)
 }
 
 // grab the value of the spork on the network, or the default
-int GetSporkValue(int nSporkID)
+int64_t GetSporkValue(int nSporkID)
 {
     int64_t r = 0;
